fixed WeightEpsilon in BOMInput

+++ Master40.DB/GeneratorModel/BillOfMaterialInput.cs
@@ -7,7 +7,7 @@ namespace Master40.DB.GeneratorModel
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
         public int EdgeWeightRoundModeId { get; set; }
-        public decimal WeightEpsilon { get; set; }
+        public double WeightEpsilon { get; set; }
         public int ApproachId { get; set; }
         public Approach Approach { get; set; }
         public EdgeWeightRoundMode EdgeWeightRoundMode { get; set; }
